refactor: rename and improve pathExist by using access function
