+++ src/fs/pathExist.mjs
@@ -1,10 +1,10 @@
-import { stat } from 'node:fs/promises'
+import { access } from 'node:fs/promises'
 
 export async function pathExist (pathUrl) {
   try {
-    const stats = await stat(pathUrl)
+    await access(pathUrl)
 
-    return !!stats
+    return true
   }
   catch (e) {
     // If the error is ENOENT (missing file/dir) so the file/dir doesn't exist
@@ -16,4 +16,3 @@ export async function pathExist (pathUrl) {
     throw e
   }
 }
-
