+++ src/fs/read.js
@@ -1,5 +1,5 @@
 import { readFile } from "node:fs/promises";
-import { pathExist } from "./fsCheck.mjs";
+import { pathExist } from "./pathExist.mjs";
 import { FS_OPERATION_FAILED, FS_UNDEFINED_ERROR } from "./fsErrorMessages.mjs";
 
 const read = async () => {
