+++ src/hash/calcHash.js
@@ -1,6 +1,6 @@
 import { stdout } from "node:process";
 import { open } from "node:fs/promises";
-import { pathExist } from "../fs/fsCheck.mjs";
+import { pathExist } from "../fs/pathExist.mjs";
 import { FS_OPERATION_FAILED } from "../fs/fsErrorMessages.mjs";
 
 const { createHash } = await import("node:crypto");
