+++ src/modules/testMjs.js
@@ -1,3 +1,3 @@
-import { unknownObject } from "./cjsToEsm.mjs";
+import { unknownObject } from "./esm.mjs";
 
 console.log("Print unknownObject in testMjs:", unknownObject);
